+++ dao/golang/dto/UserDTO.go
@@ -2,7 +2,7 @@ package dto
 
 // UserDTO - Struct representing User data
 type UserDTO struct {
-	ID   int
+	ID   float64
 	Name string
-	Age  int
+	Age  float64
 }
